+++ CHANGELOG.md
@@ -13,6 +13,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 * **Security**: in case of vulnerabilities.
 
 ## [unreleased]
+### Added
+- Add function for getting interface flags.
 
 
 ## [0.7.0] - 2025-09-12
@@ -126,4 +128,3 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 ## [0.1.0] - 2017-12-20
 ### Added
 - Initial functionality able to control most parts of the PF firewall on macOS
-